Fix TargeSubtargetInfo
In b6c22a4e the PN field was added to TargetSubtargetInfo, but this
was probably not noticed in a rebase of the book's codebase. This adds
a placeholder for this. Additionally, I've added SelectionDAG to the
link components for this target -- this appears to be necessary for me
to get the build to work.